% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggResponseSurface.R
\name{ggResponseSurface}
\alias{ggResponseSurface}
\title{Climate Response Surface}
\usage{
ggResponseSurface(data = NULL, x.specs = list(name = NULL, label =
  NULL), y.specs = list(name = NULL, label = NULL), z.specs = list(name
  = NULL, label = NULL, bins.up = NULL, bins.low = NULL, bins.mid = NULL,
  dir = "+"), display.center.value = TRUE, plot.title = NULL,
  verbose = FALSE, grid.color = NA, interpolate.grid = TRUE,
  resolution = NULL, display.contours = TRUE,
  contour.color = "gray90")
}
\arguments{
\item{data}{data frame object with x, y, and z columns}

\item{x.specs}{list of parameters specified for x dimension}

\item{y.specs}{list of parameters specified for y dimension}

\item{z.specs}{list of parameters specified for z dimension}

\item{verbose}{}
}
\description{
Climate Response Surface
}
