% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggWavelet.R
\name{ggWaveletSpectra}
\alias{ggWaveletSpectra}
\title{ggplot wrapper for wavelet spectra plot}
\usage{
ggWaveletSpectra(period, sig, obs, sim = NULL)
}
\arguments{
\item{period}{fourier periods}

\item{sig}{sigicance level}

\item{obs}{observed global wavelet spectra}

\item{sim}{simulated global wavelet spectra  default: NULL}
}
\value{
global wavelet spectra plot
}
\description{
FUNCTION_DESCRIPTION
}
\details{
place holder
}
