% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggResponseSurface.R
\name{ggResponseSurface}
\alias{ggResponseSurface}
\title{Climate response surface in ggplot}
\usage{
ggResponseSurface(data = NULL, tavg.col = "tavg", prcp.col = "prcp",
  metric.col = NULL, tavg.breaks = NULL, prcp.breaks = NULL,
  bin.pass.range = NULL, bin.fail.range = NULL, bin.threshold = NULL,
  metric.dir = "increasing", display.center.value = TRUE)
}
\arguments{
\item{data}{data frame containing climate response surfaca data}

\item{tavg.col}{name of the temperature changes column, Default: 'tavg'}

\item{prcp.col}{name of the precipitation changes column, Default: 'prcp'}

\item{metric.col}{name of the metric column, Default: NULL}

\item{tavg.breaks}{temperature change axis breaks, Default: NULL}

\item{prcp.breaks}{precipitation change axis breaks, Default: NULL}

\item{bin.pass.range}{placeholder}

\item{bin.fail.range}{placeholder}

\item{bin.threshold}{placeholder}

\item{metric.dir}{placeholder}

\item{display.center.value}{placeholder}
}
\value{
returns a ggplot2 object
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
